Pull Itcl3 change from https://core.tcl-lang.org/itcl/info/aa989cff57a23f15

+++ src/other/ext/itcl3/generic/itcl_cmds.c
@@ -866,7 +866,7 @@ Itcl_ProtectionCmd(clientData, interp, objc, objv)
     int objc;                /* number of arguments */
     Tcl_Obj *CONST objv[];   /* argument objects */
 {
-    int pLevel = (int)clientData;
+    int pLevel = (int)(size_t)clientData;
 
     int result;
     int oldLevel;
